pyproject.toml
[project]
name = "carbon-credit-marketplace"
version = "0.1.0"
description = "Decentralized Carbon Credit Marketplace"
authors = [
    {name = "1Touch-dev",email = "abhishekk@kyma.world"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "eth-ape (>=0.8.42,<0.9.0)",
    "reflex (<0.8.0)",
    "ape-solidity (>=0.8.5,<0.9.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
